refactor: add like to store when user isn't login

+++ components/artist/songs.tsx
@@ -9,11 +9,13 @@ import {useAppDispatch, useAppSelector} from "@/hooks";
 import {addSongLike, getSongsLiked, removeSongLike} from "@/lib/songsPupolar";
 import {removeLike, setLike} from "@/redux/reducers/songs-popular";
 import {useEffect} from "react";
+import useUser from "@/hooks/auth";
 
 const Songs = ({songs}) => {
     const songsLiked = useAppSelector(state => state.songPopular)
     const dispatch = useAppDispatch()
     const {setSongSrc, setSongPlay} = useGlobalContext()
+    const {auth} = useUser()
 
     const existSongLike = (songId) => songsLiked.songs.find((id) => id === songId)
 
@@ -25,6 +27,9 @@ const Songs = ({songs}) => {
     const songLike = async (songId: number) => {
         const existLike = existSongLike(songId)
 
+        if (!auth)
+            return setLike(songId)
+
         if (existLike)
             return await removeSongLike(songId).then((res) => res?.status === 200 && dispatch(removeLike(songId)))
         else
@@ -32,7 +37,8 @@ const Songs = ({songs}) => {
     }
 
     const getSongLiked = async () => {
-        await getSongsLiked().then((res) => res?.status === 200 && res.data.map(({id}) => dispatch(setLike(id))))
+        if (auth)
+            await getSongsLiked().then((res) => res?.status === 200 && res.data.map(({id}) => dispatch(setLike(id))))
     }
 
     useEffect(() => {
